+++ CHANGELOG.md
@@ -31,12 +31,14 @@ The only exceptions are the "loop" content type which still uses the "#" opening
 
 ### Added
 
+- Image plugin.
 - Template plugins can be async.
 
 ### Changed
 
+- Strongly typed `TemplateData`.
 - **BREAKING**: `RawXmlPlugin` requires data of the form `{ _type: 'rawXml', xml: string }`.
-- **BREAKING** `TemplateCompiler.compile` returns a promise.
+- **BREAKING** `TemplateCompiler.compile` (advanced api) returns a promise.
 
 ### Removed